# src/Zebra_Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Report Parser for RMF Monitor I & III to JSON",
    "version": "1.0.0",
    "title": "RMF Parsing Engine (ZEBRA)"
  },
  "host": "localhost:3090",
  "basePath": "/",
  "tags": [
    {
      "name": "RMF Monitor III",
      "description": "Convert RMF Monitor III Reports to JSON"
    },
    {
      "name": "RMF Monitor I",
      "description": "Convert RMF Monitor I Reports to JSON"
    }
  ],
  "schemes": [
    "http"
  ],
  "produces": ["application/json;charset=UTF-8"],
  "paths": {
    "/v1/{lpar}/rmf3/{report}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III Report by title only",
        "operationId": "A",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "CPC",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3/{report}?parm={value}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III CPC Report by title and caption parameter",
        "operationId": "B",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "CPC",
            "format": "string"
          },
          {
            "name": "value",
            "in": "path",
            "description": "Caption parameter to retrieve from RMF III Report(if Available)",
            "required": true,
            "type": "string",
            "default": "CPCHCMSU",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3/{report}?lpar_parms={value}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III CPC Report by title and lpar name",
        "operationId": "C",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "CPC",
            "format": "string"
          },
          {
            "name": "value",
            "in": "path",
            "description": "LPAR name to retrieve from RMF III CPC Report",
            "required": true,
            "type": "string",
            "default": "VIRPT",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3/{report}?parm={value1}&lpar_parms={value2}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III CPC Report by title, caption and Lpar name",
        "operationId": "D",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "CPC",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "Caption parameter to retrieve from RMF III Report(if Available)",
            "required": true,
            "type": "string",
            "default": "CPCHCMSU",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "LPAR name to retrieve from RMF III CPC Report",
            "required": true,
            "type": "string",
            "default": "VIRPT",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3/{report}?job={value}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III PROC and USAGE Reports by title and job name",
        "operationId": "E",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "PROC",
            "format": "string"
          },
          {
            "name": "value",
            "in": "path",
            "description": "Job name to retrieve from RMF III PROC and USAGE Report",
            "required": true,
            "type": "string",
            "default": "VIRPT",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3/{report}?resource={resource}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF III SYSINFO Report by title/Reports not accessible through gpm",
        "operationId": "F",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "SYSSUM",
            "format": "string"
          },
          {
            "name": "resource",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": ",,SYSPLEX",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3?id={id}&resource={resource}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF Metrics by ID and resource value",
        "operationId": "G1",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "LIST",
            "format": "string"
          },
          {
            "name": "resource",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": ",,SYSPLEX",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmf3?id={id}": {
      "get": {
        "tags": [
          "RMF Monitor III"
        ],
        "summary": "Get RMF Metrics by ID value",
        "operationId": "H1",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF III Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "8D0160",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I  Report by title and date parameter",
        "operationId": "G",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "CPU",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&SvcCls={value3}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I  Report by title, date and Service Class parameter",
        "operationId": "H",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Service Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "STCHIGH",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&SvcCls={value3}&Time={value4}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I Workload Report Service class information at a specific time",
        "operationId": "I",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Service Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "STCHIGH",
            "format": "string"
          },
          {
            "name": "value4",
            "in": "path",
            "description": "Exact time of Service Class Information to retrieve",
            "required": true,
            "type": "string",
            "default": "00.30.00",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&SvcCls={value3}&duration={value4}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I Workload Report Service class information at for the specified duration",
        "operationId": "J",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Service Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "STCHIGH",
            "format": "string"
          },
          {
            "name": "value4",
            "in": "path",
            "description": "The start and end time(range) of Service Class Information to retrieve",
            "required": true,
            "type": "string",
            "default": "00.30.00,03.00.00",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&Wlkd={value3}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I  Report by title, date and Workload class parameter",
        "operationId": "K",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Workload Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "TSO",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&Wlkd={value3}&Time={value4}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I workload class information at a specific time",
        "operationId": "L",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Service Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "TSO",
            "format": "string"
          },
          {
            "name": "value4",
            "in": "path",
            "description": "Exact time of Workload Class Information to retrieve",
            "required": true,
            "type": "string",
            "default": "00.30.00",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/{lpar}/rmfpp/{report}?start={value1}&end={value2}&Wlkd={value3}&duration={value4}": {
      "get": {
        "tags": [
          "RMF Monitor I"
        ],
        "summary": "Get RMF I Workload class information at for the specified duration",
        "operationId": "M",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lpar",
            "in": "path",
            "description": "LPAR from which to retrieve RMF I Report",
            "required": true,
            "type": "string",
            "default": "RPRT",
            "format": "string"
          },
          {
            "name": "report",
            "in": "path",
            "description": "Report title",
            "required": true,
            "type": "string",
            "default": "WLMGL",
            "format": "string"
          },
          {
            "name": "value1",
            "in": "path",
            "description": "start date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value2",
            "in": "path",
            "description": "end date parameter to retrieve from RMF I Report(if Available)",
            "required": true,
            "type": "string",
            "default": "2021-07-12",
            "format": "string"
          },
          {
            "name": "value3",
            "in": "path",
            "description": "Service Class Name to retrieve",
            "required": true,
            "type": "string",
            "default": "TSO",
            "format": "string"
          },
          {
            "name": "value4",
            "in": "path",
            "description": "The start and end time(range) of Workload Class Information to retrieve",
            "required": true,
            "type": "string",
            "default": "00.30.00,03.00.00",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "JSON"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  }
}